+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### bugfix
+- Update E2E tests to work with Minikube v1.39 @dbudziwojski [#757](https://github.com/newrelic/k8s-metadata-injection/pull/757)
+
 ### enhancement
 - Add support for Kubernetes 1.36, remove support for 1.31 [#755](https://github.com/newrelic/k8s-metadata-injection/pull/755)